build: Remove dnl lines with m4 itself
Instead of a non-optimally-portable regex, use m4 itself to remove dnl
lines. Plain m4 without any autoconf extensions will remove dnl, but not
expand any of the autoconf stuff. This is because autoconf undefines all
of the built-in m4 macros and defines new ones that plain m4 is unaware
of.

https://bugzilla.gnome.org/show_bug.cgi?id=774299